# jkg-api-spec.yaml
openapi: 3.0.3
info:
  title: JKG API
  description: The JKG API provides REST endpoints that return results from queries against a context instance of the [Unified Biomedical Knowledge Graph (UBKG)] in JSON Knowledge Graph (JKG) format To prevent exceeding service timeouts or payload limits, some endpoints are throttled in both execution time and in the size of its response payload--e.g., to a maximum execution time of 15 seconds and a maximum payload of 9 MB.

  version: "1.0.0"
  contact:
    name: GitHub repository
    url: https://github.com/x-atlas-consortia/jkg-api
servers:
  - url: https://jkg.api.xconsortia.org
    description: Base context instance
...
paths:
...
  /codes/{code_id}/concepts:
    get:
      operationId: codes_code_id_concepts_get
      summary: Returns the set of Concept nodes that link to the specified Code node. [A Code node can have CODE relationships with (links to) multiple Concept nodes. For example, the Code with identifier SNOMEDCT_US:79962008 links to a set of Concepts that includes C0014863 and C1442913.]
      parameters:
        - name: code_id
          in: path
          required: true
          description: The identifier for the Code node, in format _SAB_:_CODE_ where _SAB_ is an identifier for the source of the code.
          schema:
            type: string
            example: SNOMEDCT_US:79962008
      responses:
        '200':
          description: An array of Concept nodes that have CODE relationships with the specified Code node
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptDetail'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - codes
...
  /concepts/{concept_id}/codes:
    get:
      operationId: concepts_concept_id_codes_get
      summary: Returns information on the set of Code nodes that link to the specified Concept, subject to constraints specified in parameters. [A Concept node can have CODE relationships with (links to) multiple Code nodes. For example, the Concept with identifier C0006142 links to a set of Code nodes that include SNOMEDCT_US:254837009, NCI:C9335 and MSH:D001943.]
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The identifier for the Concept (also known as the Concept Unique Identifier, or CUI)
          schema:
            type: string
            example: UMLS:C0678222
        - name: sab
          in: query
          required: false
          description: The set of sources (SABs) to which to limit the set of Codes that link to the specified Concept. Can be either a list of values delimited with commas (e.g., ?sab=SAB1,SAB2,SAB3) or with individual key-value pairs (e.g., ?sab=SAB1&sab=SAB2&SAB=SAB3).
          schema:
            type: array
            uniqueItems: true
            items:
              type: string
              example: MEDLINEPLUS,MSH,MEDCIN
            default: []
      responses:
        '200':
          description: An array of code_ids in format _SAB_:_CODE for the Codes that link to the specified Concept, subject to constraints specified in parameters.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptCodes'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - concepts
  /concepts/{concept_id}/concepts:
    get:
      operationId: concepts_concept_id_concepts_get
      summary: Returns the set of Concepts that have direct relationships with the specified Concept.
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The concept identifier (also known as Concept Unique Identifier, or CUI)
          schema:
            type: string
            example: UMLS:C4722518
      responses:
        '200':
          description: An array of information on Concepts that have relationships with the specified Concept
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SabRelationshipConceptTerm'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - concepts
  /concepts/{concept_id}/definitions:
    get:
      operationId: concepts_concept_id_definitions_get
      summary: Returns the set of Definitions linked to a Concept. [A Definition is usually defined for a Code by a source (SAB), not a Concept; however, in the UBKG, the Definition is associated with the Code's linked Concept. Thus, a Concept may have multiple Definitions.]
      parameters:
        - name: concept_id
          description: Identifier (also known as a Concept Unique Identifier) for a Concept.
          in: path
          required: true
          schema:
            type: string
            example: UMLS:C0006142
      responses:
        '200':
          description: An array of Definitions associated with the Concept.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SabDefinition'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - concepts
  /concepts/{concept_id}/nodeobjects:
    get:
      operationId: concepts_identifier_nodes_get
      summary: Returns a set of "Concept Node objects" (or "concept subgraphs") that "match" the identifier (concept_id). A concept node in the UBKG is the origin of a "concept subgraph" that links a Concept node to a set of Code, Term, Definition, and Semantic Type nodes. A "Concept Node object" represents this subgraph as a set of Concept properties--i.e., all of the Concept's linked Codes, terms, definitions, and semantic types. A "match" is an exact match between an identifier and a text-based property in a Concept Node object. Because an identifier may match properties in more than one Concept Node object, the endpoint can return multiple Concept Node objects.
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The string to match against properties related to Concepts. Related properties include the preferred term for the Concept; the CUI for the Concept; the CodeID for a Code linked to the Concept; and terms for a Code linked to a Concept.
          schema:
            type: string
            example: Cells
      responses:
        '200':
          description: The set of Concept node objects that match the identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptNodes'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - concepts
...
  /terms/{term_id}/concepts:
    get:
      operationId: terms_term_id_concepts_get
      summary: Returns a list of Concept nodes with preferred terms that exactly match the specified string.
      parameters:
        - name: term_id
          in: path
          required: true
          description: The string to match. Should be URL-encoded for spaces (%20).
          schema:
            type: string
            example: Breast cancer
      responses:
        '200':
          description: An array of Concepts with preferred terms that exactly match the specified string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TermConcept'
        '303':
          description: The response size exceeds a limit set by the server. The response content is stored in a file in a AWS S3 bucket. The returned response includes a URL that can be used to download the stored file.
        '400':
          description: Invalid parameter name or value
        '403':
          description: The response size exceeds a limit set by the server.
        '404':
          description: No response for the specified parameters.
        '504':
          description: The query run time exceeded a timeout set by the server.
        '5XX':
          description: Unknown error
      tags:
        - terms
...
components:
  schemas:
...
    ConceptDetail:    # Schema name
      type: object
      description: Information on a Concept node.
      properties:
        concept:
          type: string
          description: Concept Unique Identifier (CUI) for the Concept node
          example: C0006142
...
    SabRelationshipConceptTerm: # Schema name
      type: object
      description: Information on a Concept that has a relationship with another Concept
      properties:
        sab:
          type: string
          description: The source of the relationship, such as a vocabulary or ontology, that asserts the relationship between concepts. Although relationships are usually described in terms of Codes, they actually exist between the Concepts that are linked to the Codes. For example, if an ontology SAB1 shows that code A has relationship r with code B, the associated concepts for A and B will have a relationship with type=r and SAB=SAB1.
          example: ICD10AM
        relationship:
          type: string
          description: the relationship type--e.g., "isa", "property_of"
          example: CHD
        concept:
          type: string
          description: the identifier for the Concept, also known as the Concept Unique Identifier (CUI)
          example: C0006826
        prefterm:
          type: string
          description: The preferred term for the Concept. Note that the preferred term for a Concept may be different than the preferred term for a Code that links to the Concept.
          example: Malignant Neoplasms
...
    ConceptCodes: #Schema name
      type: object
      description: The set of Code node that link to a Concept node
      properties:
        codes:
          type: array
          items:
            type: string
            description: a CodeID, in format _SAB_:_CODE_
            example: "CHV:0000052645"
...